changes space reclaim to long

+++ Project1/src/peer/Peer.java
@@ -131,7 +131,7 @@ public class Peer implements RemoteInterface {
         return "RESPONSE";
     }
 
-    public String reclaim(int spaceReclaim) {
+    public String reclaim(long spaceReclaim) {
         System.out.println("RECLAIM SERVICE -> DISK SPACE RECLAIM = " + spaceReclaim);
         // TODO:
 
@@ -141,6 +141,19 @@ public class Peer implements RemoteInterface {
     public String state() {
         System.out.println("STATE SERVICE");
         // TODO:
+
+//        - For each file whose backup it has initiated:
+//              The file pathname
+//              The backup service id of the file
+//              The desired replication degree
+//              For each chunk of the file:
+//                  Its id
+//                  Its perceived replication degree
+//        - For each chunk it stores:
+//              Its id
+//              Its size (in KBytes)
+//              Its perceived replication degree
+//        - The peer's storage capacity, i.e. the maximum amount of disk space that can be used to store chunks, and the amount of storage (both in KBytes) used to backup the chunks.
 
         return "RESPONSE";
     }
